move fetchSignature to 'core'
