+++ apps/web/app/seed/page.tsx
@@ -10,44 +10,8 @@ import {
   getPresetOptions,
   RandomizeRom,
   vanilla,
+  fetchSignature,
 } from "core";
-
-// These signatures are taken from:
-// https://github.com/dashrando/dash-template-asm/blob/main/src/fileselect/gameoptions.asm#L85-L117
-const SIGNATURE_VALUES = [
-  "GEEMER  ",
-  "RIPPER  ",
-  "ATOMIC  ",
-  "POWAMP  ",
-  "SCISER  ",
-  "NAMIHE  ",
-  "PUROMI  ",
-  "ALCOON  ",
-  "BEETOM  ",
-  "OWTCH   ",
-  "ZEBBO   ",
-  "ZEELA   ",
-  "HOLTZ   ",
-  "VIOLA   ",
-  "WAVER   ",
-  "RINKA   ",
-  "BOYON   ",
-  "CHOOT   ",
-  "KAGO    ",
-  "SKREE   ",
-  "COVERN  ",
-  "EVIR    ",
-  "TATORI  ",
-  "OUM     ",
-  "PUYO    ",
-  "YARD    ",
-  "ZOA     ",
-  "FUNE    ",
-  "GAMET   ",
-  "GERUTA  ",
-  "SOVA    ",
-  "BULL    ",
-];
 
 function getSeedOpts(): {
   num: number;
@@ -73,17 +37,6 @@ function getSeedOpts(): {
 
 function downloadFile(data: any, name: string) {
   saveAs(new Blob([data]), name);
-}
-
-function fetchSignature(data: any) {
-  // the signature is stored in 4 bytes at 0x2f8000 - 0x2f8003
-  // use bit mask of 0x1f to get the index in the signatures array
-  // then trim the string to remove the extra spaces
-  const mask = 0x1f;
-  const addresses = [0x2f8000, 0x2f8001, 0x2f8002, 0x2f8003]
-    .map((addr) => data[addr] & mask)
-    .map((index) => SIGNATURE_VALUES[index].trim());
-  return addresses.join(" ");
 }
 
 export default function SeedPage() {
